// client/src/pages/admins/helpers/_schemas.tsx
export const EditContactSchema = yup.object().shape({
    company_contact: yup.string().min(3, "Longitud minima 3 digitos").required("Campo obligatorio"),
    company_email: yup.string().email("Formato invalido").min(3, "Longitud minima 3 digitos").required("Campo obligatorio"),
    company_phone: yup.string().min(3, "Longitud minima 3 digitos").required("Campo obligatorio"),
    company_job: yup.string().min(3, "Longitud minima 3 digitos").required("Campo obligatorio"),
})

export const initialValuesContact = {
    company_contact: "",
    company_email: "",
    company_phone: "",
    company_job: "",
}